bug: Handle TypeError in Django migration
CI Lesson: #

Developing Django /I Think Therefore I Blog/ Models part 1/
Creating the database Step 17-18

The current migration process in our Django project is encountering
a TypeError. This error occurs when we attempt to run the
`python3 manage.py migrate` command. The issue is traced back to the
`settings.py` file, specifically on line 95.

The error message "TypeError: a bytes-like object is required, not
'str’" suggests that the `dj_database_url.parse` method is expecting
a bytes-like object, but it is receiving a string from `os.environ.get`.

The current implementation of the code is as follows:

```python
DATABASES = {
    'default': dj_database_url.parse(os.environ.get("postgres://auxfhsvc:vnjyWR-g8ujQQ8devHsWH4uq2d9IBOEk@snuffleupagus.db.elephantsql.com/auxfhsvc"))
}

```
On branch main
Your branch is up to date with 'origin/main'.
Changes to be committed:
	new file:   .gitpod.yml
	modified:   codestar/settings.py
	modified:   requirements.txt