fix(database): Corrected database configuration
An error occurred when running the Django migration due to a misconfigured DATABASES setting in the [settings.py](http://settings.py/) file. The ENGINE value was missing, causing the migration to fail. The [settings.py](http://settings.py/) file has been updated with the correct ENGINE value.

Resolves: Database configuration issue.

On branch main
Your branch is up to date with 'origin/main'.
Changes to be committed:
	new file:   .tours/3-creating-the-database.tour
	modified:   codestar/settings.py