fix(database): Corrected DATABASES setting typo
The error was due to a misnaming in the DATABASES setting in the Django configuration. The setting was erroneously named `DATABASE`. Compared the user's code with the source code from the walkthrough and identified the discrepancy. Suggested using tools like [diffchecker.com](http://diffchecker.com/) for spotting such differences in the future.

Resolves: Database setting naming issue.

On branch main
Your branch is up to date with 'origin/main'.
Changes to be committed:
	modified:   .tours/3-creating-the-database.tour
	modified:   codestar/settings.py